+++ src/main/java/com/library/entities/Vente.java
@@ -29,7 +29,7 @@ public class Vente implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	private int numeroVente;
+	private long numeroVente;
 
 	//@DateTimeFormat(pattern = "yyyy-MM-dd")
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "GMT")
@@ -47,7 +47,7 @@ public class Vente implements Serializable {
 	public Vente() {
 	}
 
-	public Vente(Long id, int numeroVente, Date dateVente, @Valid List<LigneVente> ligneVentes, double totalVente, String status) {
+	public Vente(Long id, long numeroVente, Date dateVente, @Valid List<LigneVente> ligneVentes, double totalVente, String status) {
 		this.id = id;
 		this.numeroVente = numeroVente;
 		this.dateVente = dateVente;
@@ -64,11 +64,11 @@ public class Vente implements Serializable {
 		this.id = id;
 	}
 
-	public int getNumeroVente() {
+	public long getNumeroVente() {
 		return numeroVente;
 	}
 
-	public void setNumeroVente(int numeroVente) {
+	public void setNumeroVente(long numeroVente) {
 		this.numeroVente = numeroVente;
 	}
