+++ src/main/java/com/library/repository/LigneVenteRepository.java
@@ -14,11 +14,11 @@ import java.util.List;
 @Repository
 public interface LigneVenteRepository extends JpaRepository<LigneVente, Long> {
 
-    List<LigneVente> findAllByNumero(int numero);
+    List<LigneVente> findAllByNumero(long numero);
 
     @Modifying
     @Query("delete from LigneVente where numero = :numero")
-    void deleteByNumero(@Param("numero") int numero);
+    void deleteByNumero(@Param("numero") long numero);
 
     @Query("select p from LigneVente p where p.produit.id =:prod")
     List<LigneVente> ListLigneVenteByProduitId(@Param("prod") Long prodId);
